+++ tooling/scripts/tests/fixtures/method_chain_wrap/09_chain_with_source_preserved_args_inline/input.java
@@ -1,9 +1,0 @@
-public class Demo
-{
-    public void run()
-    {
-        Class<Demo> cls = Demo.class;
-        String url = cls.getResource(
-                cls.getSimpleName() + ".class").toString();
-    }
-}